+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.js
@@ -13,11 +13,17 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { useResizeDetector } from 'react-resize-detector';
 import { moderate02 } from '@carbon/motion';
+
+import { getDevtoolsProps } from '../../global/js/utils/devtools';
+
+import {
+  allPropTypes,
+  deprecateProp,
+} from '../../global/js/utils/props-helper';
+
 import wrapFocus from '../../global/js/utils/wrapFocus';
 import { pkg } from '../../settings';
-import { allPropTypes } from '../../global/js/utils/props-helper';
 import { SIDE_PANEL_SIZES } from './constants';
-import { deprecateProp } from '../../global/js/utils/props-helper';
 
 // Carbon and package components we use.
 import { Button } from 'carbon-components-react';
@@ -197,7 +203,8 @@ export let SidePanel = React.forwardRef(
             : sidePanelSubtitleElementHeight;
         sidePanelSubtitleElementHeight =
           sidePanelSubtitleElementHeight < 0
-            ? 16
+            ? sidePanelScrollArea?.scrollHeight -
+              sidePanelScrollArea?.clientHeight
             : sidePanelSubtitleElementHeight;
         /* istanbul ignore next */
         sidePanelScrollArea &&
@@ -373,6 +380,12 @@ export let SidePanel = React.forwardRef(
           onRequestClose();
         }
       };
+      const bodyElement = document.body;
+      if (includeOverlay && open) {
+        bodyElement.style.overflow = 'hidden';
+      } else if (includeOverlay && !open) {
+        bodyElement.style.overflow = 'initial';
+      }
       if (includeOverlay) {
         document.addEventListener('click', handleOutsideClick);
       }
@@ -498,7 +511,7 @@ export let SidePanel = React.forwardRef(
           actionToolbarButtons && actionToolbarButtons.length,
         [`${blockClass}__container-without-overlay`]:
           !includeOverlay && !slideIn,
-        [`${blockClass}__container-is-animating`]: !animationComplete,
+        [`${blockClass}__container-is-animating`]: !animationComplete || !open,
       },
     ]);
 
@@ -509,9 +522,10 @@ export let SidePanel = React.forwardRef(
             [`${blockClass}__on-detail-step`]: currentStep > 0,
             [`${blockClass}__on-detail-step-without-title`]:
               currentStep > 0 && !title,
-            [`${blockClass}__title-container--no-animation`]: !animateTitle,
+            [`${blockClass}__title-container--no-title-animation`]:
+              !animateTitle,
             [`${blockClass}__title-container-is-animating`]:
-              !animationComplete && animateTitle,
+              !animationComplete || !open,
             [`${blockClass}__title-container-without-title`]: !title,
           })}>
           {currentStep > 0 && (
@@ -584,9 +598,8 @@ export let SidePanel = React.forwardRef(
                 onClick={(event) =>
                   action.onClick
                     ? action.onClick(event)
-                    : action.onActionToolbarButtonClick
-                    ? action.onActionToolbarButtonClick(event)
-                    : null
+                    : action.onActionToolbarButtonClick &&
+                      action.onActionToolbarButtonClick(event)
                 }>
                 {action.leading && action.label}
               </Button>
@@ -629,10 +642,8 @@ export let SidePanel = React.forwardRef(
       shouldRender && (
         <>
           <div
-            {
-              // Pass through any other property values as HTML attributes.
-              ...rest
-            }
+            {...getDevtoolsProps(componentName)}
+            {...rest}
             id={`${blockClass}-outer`}
             className={mainPanelClassNames}
             style={{
@@ -664,6 +675,8 @@ export let SidePanel = React.forwardRef(
               ref={sidePanelInnerRef}
               className={cx(`${blockClass}__inner-content`, {
                 [`${blockClass}__static-inner-content`]: !animateTitle,
+                [`${blockClass}__inner-content-with-actions`]:
+                  actions && actions.length,
               })}>
               {animateTitle && renderHeader()}
               <div className={`${blockClass}__body-content`}>{children}</div>